Allow specifying multiple sources for shader program
It also extracts shader language version. It is now added at runtime to
all shaders identically.

This might be a problem in the future but I just wanted a simple way to
reuse definitions in shader language so for now just joining multiple
files.
In the future we could add out own #include(../path) directive to shaders that we
would search for and replace it with specified file contents.